coap: support request re-sending in CoapClient
This command brings support for re-sending request in case of timeout
while waiting for response. FobnailClient does not need to handle
timeouts itself anymore. Also this fixes a problem when the same request
when re-sent has a new token each time. This can be mishandled by server
as a separate request (instead of being ignored as duplicate).