pal_pc: improve Trussed support
Allow to create multiple Trussed clients, use a custom syscall
implementation (still dummy). Originally this was done to allow creating
multiple Trussed clients, to have a separate class for managing (loading
and unloading) keys which was ment to solve problem with panics due to
multiple borrowing of trussed client used in FobnailClient. But a key is
valid only in context of client which created that key.