pal_nrf: erase persistent storage when reflashing firmware
When flashing, cargo-embed erases only sectors that will be overridden,
leaving persistent storage. This behaviour is not always ideal. Ideally
it should be possible to control using command line switches that
behaviour but cargo-embed does not support it, for now disable NOLOAD in
linker script so that cargo-embed always overrides storage.