Harden AIK verification.
Previous approach for verifying AIK had was not secure, an attacker
could replace AIK because we were not checking whether received AIK
matches with it's name used for credential activation. Now, instead of
relying on AIK name provided from atteter we compute it ourselves. If
AIK has been tampered with, hashes will be different and credential
activation will fail.
Another advantage is that we can confirm that the received key has
proper attributes set, as required for AIK keys.