Update Trussed.
Use Trussed version with modified MAX_MESSAGE_LENGTH. We need this
because some certificates are too big. Since we are using lfs tool for
uploading certificates we can write files bigger than that but we cannot
read them, Trussed will read only the first MAX_MESSAGE_LENGTH bytes
from file. In case of X.509 this may look as if certificate is corrupted
or the parser broken.