certmgr: support veryfing of PO certificates
This commit introduces two verification modes, one is for verifying EK
certificates and it retains previous behaviour, new mode is for PO
(Platform Owner) certificates and it has applies different verification
policy. Additionally, from now on, certificates may be temporarily
injected (are gone with certmgr instance), this is used for PO chain
verification.